moved camera view scripts into proper folders and added namespace and copyright
