--- 3DAmsterdam/Assets/Camera2DFrustum.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 60beef4fb16356a41a6ba254801b6a77
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
