+++ 3DAmsterdam/Assets/Netherlands3D/Scripts/Interface/UIQuad.cs
@@ -1,3 +1,20 @@
+/*
+*  Copyright (C) X Gemeente
+*                X Amsterdam
+*                X Economic Services Departments
+*
+*  Licensed under the EUPL, Version 1.2 or later (the "License");
+*  You may not use this work except in compliance with the License.
+*  You may obtain a copy of the License at:
+*
+*    https://github.com/Amsterdam/3DAmsterdam/blob/master/LICENSE.txt
+*
+*  Unless required by applicable law or agreed to in writing, software
+*  distributed under the License is distributed on an "AS IS" basis,
+*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+*  implied. See the License for the specific language governing
+*  permissions and limitations under the License.
+*/
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
@@ -22,12 +39,16 @@ namespace Netherlands3D.Interface
         public void Redraw()
         {
             SetVerticesDirty();
-            
+            CalculateBounds();
+        }
+        private void CalculateBounds()
+        {
             //Make sure our graphic width/height is set to the max distance of our verts, so culling works properly
-            /*rectTransform.sizeDelta = new Vector2(
-                Mathf.Abs(Mathf.Max(QuadVertices[0].x, QuadVertices[1].x, QuadVertices[2].x, QuadVertices[3].x)),
-                Mathf.Abs(Mathf.Max(QuadVertices[0].y, QuadVertices[1].y, QuadVertices[2].y, QuadVertices[3].y))
-            );*/
+            //Now based from centered origin
+            rectTransform.sizeDelta = new Vector2(
+                Mathf.Abs(2 * Mathf.Max(QuadVertices[0].x, QuadVertices[1].x, QuadVertices[2].x, QuadVertices[3].x)),
+                Mathf.Abs(2 * Mathf.Min(QuadVertices[0].y, QuadVertices[1].y, QuadVertices[2].y, QuadVertices[3].y))
+            );
         }
 
         private void UpdateQuadVertexList()
